Move the presentation edit component to better locations
